+++ src/core/bc.h
@@ -81,15 +81,18 @@ struct au_bc_storage {
     int num_args;
 };
 
+/// [func] Initializes an au_bc_storage instance
+/// @param bc_storage instance to be initialized
 void au_bc_storage_init(struct au_bc_storage *bc_storage);
+
+/// [func] Deinitializes an au_bc_storage instance
+/// @param bc_storage instance to be deinitialized
 void au_bc_storage_del(struct au_bc_storage *bc_storage);
 
 struct au_program_data;
 
-/// Debugs an bytecode storage container
+/// [func] Debugs an bytecode storage container
 /// @param bcs the bytecode storage
 /// @param data program data
 void au_bc_dbg(const struct au_bc_storage *bcs,
                const struct au_program_data *data);
-
-struct au_program;