+++ src/core/rt/au_string.c
@@ -3,10 +3,9 @@
 //
 // Licensed under Apache License v2.0 with Runtime Library Exception
 // See LICENSE.txt for license information
-#include <string.h>
-
 #ifdef AU_IS_INTERPRETER
 #include "au_string.h"
+#include <string.h>
 #endif
 
 struct au_string *au_string_from_const(const char *s, size_t len) {