+++ src/platform/mmap.h
@@ -15,11 +15,13 @@ struct au_mmap_info {
 #endif
 };
 
-/// Loads a file into memory and stores a reference into a au_mmap_info
-/// instance
+/// [func] Loads a file into memory and stores a reference
+/// into a au_mmap_info instance
 /// @param path path of file
 /// @param info info
 /// @return 1 if success, 0 if errored
 int au_mmap_read(const char *path, struct au_mmap_info *info);
 
-void au_mmap_close(struct au_mmap_info *info);
+/// [func] Deinitializes an au_mmap_info instance
+/// @param info instance to be deinitialized
+void au_mmap_del(struct au_mmap_info *info);