+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# Version 0.0.x [pre-α]
+# Version 0.1 [pre-α]
+
+  * Floating point support
+  * Added standard library functions `bool`, `input`, `int`, `str`
+  * Implemented FNV hash
+
+# Version 0.0 [pre-α]
 
   * Core functionality including:
     * Basic data structures: Integers, strings and booleans